Code refactoring, decoupled the application from text processing.
